+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Core: Switched to a requests.session object to enable persistent connection recycling.
 - Core: Provide options to disable pagination and prevent api result's JSON parsing (in association with bulk_export).
 
-### Bug Fixes
+### Fixed
 
 - Core: do not prepend the base_url to the uri of an api_call if a protocol is already included (i.e. uri is already fully-formed).